Add files via upload
The below Visualization report is about the NUMBER OF EMPLOYEES BY AGE GROUP of a company data. In which how many number of employees are working based on age group.
The application which I am using for my visualization reports and dashboards is TABLEAU which is very helpful for report creation and Dashboard generation about the data for analytics purpose.